Major rebase of code
Basically what this branch deserves its name for
