+++ config/keys.conf
@@ -41,7 +41,7 @@ Shift+h:    left_page
 Shift+j:    down_page
 Shift+k:    up_page
 Shift+l:    right_page
-Shift+o:    library_focus
+Shift+o:    focus_library
 g:          first
 Shift+g:    last
 
@@ -67,7 +67,7 @@ c:          manipulate
 
 [LIBRARY] ######################################################################
 # Keybindings that only apply in the library
-Shift+o:    library_unfocus
+Shift+o:    unfocus_library
 t:          set show_hidden!
 h:          left_lib
 j:          down_lib
