+++ vimiv/parser.py
@@ -14,25 +14,25 @@ def get_args():
 
     parser.add_argument("-b", "--bar", action="store_true", dest="sbar",
                         help="display statusbar")
-    parser.add_argument("-B", "--no-bar", action="store_false", dest="sbar",
+    parser.add_argument("-B", "--no-bar", action="store_true", dest="hide_sbar",
                         help="hide statusbar")
     parser.add_argument("-l", "--library", action="store_true",
                         dest="library", help="display library")
-    parser.add_argument("-L", "--no-library", action="store_false",
-                        dest="library", help="don't display library")
+    parser.add_argument("-L", "--no-library", action="store_true",
+                        dest="hide_library", help="don't display library")
     parser.add_argument("-f", "--fullscreen", action="store_true",
                         dest="fullscreen", help="start in fullscreen")
-    parser.add_argument("-F", "--no-fullscreen", action="store_false",
-                        dest="fullscreen", help="don't start in fullscreen")
+    parser.add_argument("-F", "--no-fullscreen", action="store_true",
+                        dest="no_fullscreen", help="don't start in fullscreen")
     parser.add_argument("-s", "--shuffle", action="store_true",
                         dest="shuffle", help="shuffle filelist")
-    parser.add_argument("-S", "--no-shuffle", action="store_false",
-                        dest="shuffle", help="don't shuffle the filelist")
+    parser.add_argument("-S", "--no-shuffle", action="store_true",
+                        dest="no_shuffle", help="don't shuffle the filelist")
     parser.add_argument("-r", "--recursive", action="store_true",
                         dest="recursive",
                         help="search given directories recursively")
-    parser.add_argument("-R", "--no-recursive", action="store_false",
-                        dest="recursive",
+    parser.add_argument("-R", "--no-recursive", action="store_true",
+                        dest="no_recursive",
                         help="don't search given directories recursively")
     parser.add_argument("--slideshow", action="store_true",
                         help="start slideshow immediately")
@@ -52,21 +52,31 @@ def parse_args(parser, settings):
     """ Parse the arguments and return the modified settings """
     args = parser.parse_args()
     if args.sbar:
-        settings["GENERAL"]["display_bar"] = args.sbar
+        settings["GENERAL"]["display_bar"] = True
+    if args.hide_sbar:
+        settings["GENERAL"]["display_bar"] = False
     if args.recursive:
-        settings["GENERAL"]["recursive"] = args.recursive
+        settings["GENERAL"]["recursive"] = True
+    if args.no_recursive:
+        settings["GENERAL"]["recursive"] = False
     if args.fullscreen:
-        settings["GENERAL"]["start_fullscreen"] = args.fullscreen
+        settings["GENERAL"]["start_fullscreen"] = True
+    if args.no_fullscreen:
+        settings["GENERAL"]["start_fullscreen"] = False
     if args.shuffle:
-        settings["GENERAL"]["shuffle"] = args.shuffle
+        settings["GENERAL"]["shuffle"] = True
+    if args.no_shuffle:
+        settings["GENERAL"]["shuffle"] = False
     if args.slideshow_delay:
         settings["GENERAL"]["slideshow_delay"] = args.slideshow_delay
     if args.slideshow:
-        settings["GENERAL"]["start_slideshow"] = args.slideshow
+        settings["GENERAL"]["start_slideshow"] = True
     if args.geometry:
         settings["GENERAL"]["geometry"] = args.geometry
     if args.library:
-        settings["LIBRARY"]["show_library"] = args.library
+        settings["LIBRARY"]["show_library"] = True
+    if args.hide_library:
+        settings["LIBRARY"]["show_library"] = False
 
     settings["GENERAL"]["start_from_desktop"] = args.desktop
     # Be able to read files from a stdin pipe
